refactor: update file structure to adhere to docs
