+++ client/app/features/map/MapSideBar.tsx
@@ -1,0 +1,7 @@
+export default function MapSideBar() {
+    return (
+        <>
+            <div></div>
+        </>
+    );
+}
