+++ client/app/routes/map2.tsx
@@ -1,0 +1,5 @@
+import MapSideBar from "../features/map/MapSideBar";
+
+export default function MapSideBarPage() {
+    return <MapSideBar />;
+}
