+++ client/package.json
@@ -16,7 +16,6 @@
     "dependencies": {
         "@react-router/node": "7.12.0",
         "@react-router/serve": "7.12.0",
-        "@types/leaflet": "^1.9.21",
         "isbot": "^5.1.31",
         "leaflet": "^1.9.4",
         "react": "^19.2.4",
@@ -31,6 +30,7 @@
         "@testing-library/jest-dom": "^6.9.1",
         "@testing-library/react": "^16.3.2",
         "@testing-library/user-event": "^14.6.1",
+        "@types/leaflet": "^1.9.21",
         "@types/node": "^22",
         "@types/react": "^19.2.7",
         "@types/react-dom": "^19.2.3",
